Fix: Return count of unique compounds loaded #314
loadCompoundCSVFile returns the number of rows read from a CSV. This only includes the rows that have either the mz or the formula information.

The issue was resolved for this function in 'databases' class in a previous commit.

This commit will change behavior of the function in 'database' class to return the number of compounds loaded after removing duplicates.

Issue: #314