Test: Remove redundant test case for compound loading #314
There was a test case added for loadCompoundsCSVFile to check whether duplicate compounds are being loaded only once if they do not have an ID.

Currently, ID is used for detection of duplicates. If ID is not provided, compound name is used. ID is auto-assigned if both are not provided in the CSV.

Unless every field of every compound is compared, this corner case cannot be covered.

Issue: #314