Fix: barplot/boxplot missing #316
Barplot, Boxplot and Isotopeplot were missing because of minor change in computation method for maxIntensity vector.This change was introduced while resolving issue #261. When a sample was not selected for peak finding, maxIntensity for this sample in a group was being assigned -1 instead of 0.

Now it's been restored to previous state and to ressolve issue #261 we are directly checking by comparing all sample and used sample ( samplesUsed ) for peak finding .

issue #316